Move retry strategy to separate package
