+++ go/common/retry/retry_strategy.go
@@ -1,13 +1,13 @@
-package clientutil
+package retry
 
 import (
 	"fmt"
 	"time"
 )
 
-// retryStrategy interface allows for flexible strategies for retrying/polling functions
-type retryStrategy interface {
-	// WaitInterval calls can be considered as marking the completion of an attempt
+// RetryStrategy interface allows for flexible strategies for retrying/polling functions, the usage should be:
+type RetryStrategy interface {
+	// NextRetryInterval calls can be considered as marking the completion of an attempt
 	NextRetryInterval() time.Duration // returns the duration to sleep before making the next attempt (may not be fixed, e.g. if strategy is to back-off)
 	Done() bool                       // returns true when caller should stop retrying
 	Summary() string                  // message to summarise usage (i.e. number of retries, time take, if it failed and why, e.g. "timed out after 120 seconds (8 attempts)"
